// packages/expo-modules-autolinking/src/types.ts
export type PackageRevision = {
  name: string;
  path: string;
  version: string;
  config?: ExpoModuleConfig;
  duplicates?: PackageRevision[];
};
// ...
export type HarmonyModuleKind =
  | 'turbo-module'
  | 'view'
  | 'ability-lifecycle'
  | 'rnoh-reuse'
  | 'rnoh-package'
  | 'har';

export interface HarmonyCppPackageConfig {
  packageClass: string;
  header: string;
  cmakeTarget: string;
  cmakePath: string;
}

export interface HarmonyEtsPackageConfig {
  packageClass: string;
  importPath: string;
  entrypoint: string;
  importKind?: 'named' | 'default';
}

export interface HarmonyLifecycleDependency {
  localName: string;
  className: string;
  importPath: string;
  appStorageKey: string;
  arkUIOverlay?: {
    className: string;
    importPath: string;
    controllerProperty: string;
  };
}

export interface HarmonyHarConfig {
  packageName: string;
  packagePath: string;
  primary?: boolean;
  version?: string;
  transform?: 'screens-content-wrapper-v1' | 'worklets-private-symbols-v2';
}

export interface RawModuleConfigHarmony {
  kind: HarmonyModuleKind;
  cpp?: HarmonyCppPackageConfig;
  ets?: HarmonyEtsPackageConfig;
  lifecycleDependencies?: HarmonyLifecycleDependency[];
  har?: HarmonyHarConfig | HarmonyHarConfig[];
  reason?: string;
}

export interface ModuleDescriptorHarmony {
  packageName: string;
  packageVersion?: string;
  packageRoot: string;
  kind: HarmonyModuleKind;
  cpp?: HarmonyCppPackageConfig;
  ets?: HarmonyEtsPackageConfig;
  lifecycleDependencies: HarmonyLifecycleDependency[];
  hars?: HarmonyHarConfig[];
  reason?: string;
}
// ...
type AndroidMavenRepositoryCredentials =
  | AndroidMavenRepositoryPasswordCredentials
  | AndroidMavenRepositoryHttpHeaderCredentials
  | AndroidMavenRepositoryAWSCredentials;

export interface AndroidMavenRepository {
  /**
   * The URL of the Maven repository.
   */
  url: string;
  /**
   * The credentials to use when accessing the Maven repository.
   * May be of type PasswordCredentials, HttpHeaderCredentials, or AWSCredentials.
   *
   * @see the authentication schemes section of [Gradle documentation](https://docs.gradle.org/current/userguide/declaring_repositories.html#sec:authentication_schemes) for more information.
   */
  credentials?: AndroidMavenRepositoryCredentials;
  /**
   * The authentication scheme to use when accessing the Maven repository.
   */
  authentication?: 'basic' | 'digest' | 'header';
}

interface ApplePod {
  name: string;
  version?: string;
  configurations?: string[];
  modular_headers?: boolean;
  source?: string;
  path?: string;
  podspec?: string;
  testspecs?: string[];
  git?: string;
  branch?: string;
  tag?: string;
  commit?: string;
}

export type ExtraDependencies = AndroidMavenRepository[] | ApplePod[];